+++ pennylane_lightning/lightning_qubit.py
@@ -15,7 +15,7 @@ r"""
 This module contains the :class:`~.LightningQubit` class, a PennyLane simulator device that
 interfaces with C++ for fast linear algebra calculations.
 """
-from typing import List
+from typing import Iterable, List
 from warnings import warn
 from os import getenv
 from itertools import islice
@@ -203,7 +203,11 @@ class LightningQubit(DefaultQubit):
         else:
             raise TypeError(f"Unsupported complex Type: {dtype}")
 
-        for o in operations:
+        skipped_ops = ["Identity"]
+        ops_iter = operations if isinstance(operations, Iterable) else list(operations)
+        ops_generator = (o for o in ops_iter if o not in skipped_ops)
+
+        for o in ops_generator:
             name = o.name.split(".")[0]  # The split is because inverse gates have .inv appended
             if _is_lightning_gate(name):
                 kernel = self._kernel_for_ops[name]