+++ tests/test_apply.py
@@ -16,6 +16,7 @@ Unit tests for the :mod:`pennylane_lightning.LightningQubit` device.
 """
 # pylint: disable=protected-access,cell-var-from-loop
 import math
+from tracemalloc import start
 
 import numpy as np
 import pennylane as qml
@@ -118,6 +119,8 @@ class TestApply:
         ),
         (qml.Hadamard, [1, 0], [1 / math.sqrt(2), 1 / math.sqrt(2)]),
         (qml.Hadamard, [1 / math.sqrt(2), -1 / math.sqrt(2)], [0, 1]),
+        (qml.Identity, [1, 0], [1, 0]),
+        (qml.Identity, [1 / math.sqrt(2), 1 / math.sqrt(2)], [1 / math.sqrt(2), 1 / math.sqrt(2)]),
     ]
 
     @pytest.mark.parametrize("operation,input,expected_output", test_data_no_parameters)
@@ -1432,6 +1435,46 @@ class TestTensorSample:
             - 2 * np.cos(theta) * np.sin(phi) * np.sin(2 * varphi)
         ) / 4
         assert np.allclose(var, expected, atol=tolerance, rtol=0)
+
+
+class TestApplyLightningMethod:
+    """Unit tests for the apply_lightning method."""
+
+    def test_noniter_identity_skipped(self, mocker, tol):
+        """Test identity operation does not perform additional computations."""
+        dev = qml.device("lightning.qubit", wires=1)
+
+        starting_state = np.array([1, 0], dtype=np.complex128)
+        op = qml.Identity(0)
+
+        spy_diagonal = mocker.spy(dev, "_apply_diagonal_unitary")
+        spy_einsum = mocker.spy(dev, "_apply_unitary_einsum")
+        spy_unitary = mocker.spy(dev, "_apply_unitary")
+
+        res = dev.apply_lightning(starting_state, op, dtype=np.complex128)
+        assert np.allclose(res, starting_state, atol=tol, rtol=0)
+
+        spy_diagonal.assert_not_called()
+        spy_einsum.assert_not_called()
+        spy_unitary.assert_not_called()
+
+    def test_iter_identity_skipped(self, mocker, tol):
+        """Test identity operation does not perform additional computations."""
+        dev = qml.device("lightning.qubit", wires=2)
+
+        starting_state = np.array([1, 0, 0, 0], dtype=np.complex128)
+        op = [qml.Identity(0), qml.Identity(1)]
+
+        spy_diagonal = mocker.spy(dev, "_apply_diagonal_unitary")
+        spy_einsum = mocker.spy(dev, "_apply_unitary_einsum")
+        spy_unitary = mocker.spy(dev, "_apply_unitary")
+
+        res = dev.apply_lightning(starting_state, op, dtype=np.complex128)
+        assert np.allclose(res, starting_state, atol=tol, rtol=0)
+
+        spy_diagonal.assert_not_called()
+        spy_einsum.assert_not_called()
+        spy_unitary.assert_not_called()
 
 
 def test_warning():